fix: fix validationSchema erorr
